feat: move cniargs type define from utils to types
